// packages/rum-core/src/rawRumEvent.types.ts
export const enum RumEventType {
  ACTION = 'action',
  ERROR = 'error',
  LONG_TASK = 'long_task',
  VIEW = 'view',
  RESOURCE = 'resource',
  VITAL = 'vital',
}